+++ app/src/main/java/com/vodovoz/app/data/vodovoz_service/model/ZAKAZ_DTO.kt
@@ -9,7 +9,7 @@ data class ZAKAZ_DTO(
     @Json(name = "BORDERCOLOR")
     val BORDERCOLOR: String?,
     @Json(name = "IDZAKAZ")
-    val IDZAKAZ: Int?,
+    val IDZAKAZ: Long?,
     @Json(name = "OPISANIE")
     val OPISANIE: String?,
     @Json(name = "PRICE")
